backend/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pydantic-settings
langgraph
langchain
faiss-cpu
langchain-community
sentence-transformers
fastembed
langchain-google-genai
